Separated ProductListAdapter into reusable components
